requirements.txt
aiomysql==0.0.21
aiosqlite==0.16.0
asynctest==0.13.0
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
httpcore==0.12.2
httpx==0.16.1
idna==3.0
iniconfig==1.1.1
iso8601==0.1.13
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
PyMySQL==0.9.3
pyparsing==2.4.7
PyPika==0.44.1
pytest==6.2.1
pytest-asyncio==0.14.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.5
rfc3986==1.4.0
rsa==4.6
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.16.19
typing-extensions==3.7.4.3
uvicorn==0.13.3
wincertstore==0.2
